refactor(chat): 抽取 formatDuration 为独立模块，收窄工具判定函数可见性
